platform-backend/C(men)/agents/learning-mode/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
openai==1.30.1
httpx==0.27.0
python-dotenv==1.0.1
azure-cosmos==4.7.0
azure-storage-blob==12.20.0